interview_ai_agent/requirements.txt
langgraph>=0.2.40
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.50.0
moviepy==1.0.3
opencv-python>=4.8.0
numpy>=1.26.0
pydantic>=2.8.0
python-dotenv>=1.0.1
requests>=2.31.0
